refactor(config): hold the view distance at the protocol's width
#27 fixed the wrap by bounding the input, and CodeQL kept reporting all three
go/incorrect-integer-conversion alerts afterwards -- correctly. The conversions
were still there; only a validator two packages away made them safe, and taint
analysis cannot see that. Three high-severity alerts that are false positives
are worse than none, because the next reader cannot tell them from the real
finding they started as.

So remove the conversions instead of guarding them. Config.ViewDistance becomes
a uint8, which is the width RequestChunkRadius.MaxChunkRadius already uses, and
strconv.ParseUint with a bitSize of 8 refuses at parse time anything that would
not survive the narrowing. The packet now takes cfg.ViewDistance directly, and
ChunkRadius widens to int32, which is always exact.

The out-of-range value is no longer representable rather than merely refused.
That holds even if someone later loosens maxViewDistance, which the #27 shape
did not.

positiveInt keeps its required max for MC_PORT and the backoff pair; those feed
range-bound values rather than narrowed ones, so nothing there needs the same
treatment.

Tests gain the 255 case: 255 and 256 now fail by different routes -- one cannot
parse into a uint8, the other parses and exceeds the ceiling -- and a change
collapsing one into the other would be invisible to a test that only tried 256.

+++ cmd/bot/main.go
@@ -159,8 +159,8 @@ func session(ctx context.Context, cfg config.Config, ts oauth2.TokenSource, log 
 	// loads nothing is indistinguishable from a healthy one until someone
 	// notices the farm has stopped.
 	if err := conn.WritePacket(&packet.RequestChunkRadius{
-		ChunkRadius:    cfg.ViewDistance,
-		MaxChunkRadius: uint8(cfg.ViewDistance),
+		ChunkRadius:    int32(cfg.ViewDistance),
+		MaxChunkRadius: cfg.ViewDistance,
 	}); err != nil {
 		return fmt.Errorf("request chunk radius: %w", err)
 	}
@@ -190,8 +190,8 @@ func session(ctx context.Context, cfg config.Config, ts oauth2.TokenSource, log 
 				// player is replaced, so it is asked for again rather than
 				// assumed to have survived.
 				if err := conn.WritePacket(&packet.RequestChunkRadius{
-					ChunkRadius:    cfg.ViewDistance,
-					MaxChunkRadius: uint8(cfg.ViewDistance),
+					ChunkRadius:    int32(cfg.ViewDistance),
+					MaxChunkRadius: cfg.ViewDistance,
 				}); err != nil {
 					return fmt.Errorf("request chunk radius after respawn: %w", err)
 				}
